FindAutoCount
An Autobiographical Number is a number N such that the first digit of N represents the count of how many zeroes are there in N, the second digit represents the count of how many ones are there in N, and so on.

You are given a function: def FindAutoCount(n):

The function accepts a string n which is a number and checks whether the number is an autobiographical number or not. If it is, an integer is returned, i.e., the count of distinct numbers in n. If not, it returns 0.

Assumption:

The input string will not be longer than 10 characters.
Input string will consist of numeric characters.
Note:

If the string is None, return 0.